move utils to rdma
