+++ csrc/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_subdirectory(engine)
-add_subdirectory(utils)
 
 add_subdirectory(ops)
