+++ csrc/engine/rdma/rdma_assignment.h
@@ -16,7 +16,7 @@
 #include <thread>
 #include <unordered_map>
 
-#include "utils/json.hpp"
+#include "json.hpp"
 
 namespace slime {
