+++ csrc/engine/rdma/rdma_config.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "utils/json.hpp"
+#include "json.hpp"
 #include "logging.h"
 
 #include <cstddef>
