+++ csrc/engine/rdma/rdma_context.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include "engine/assignment.h"
+#include "engine/rdma/affinity.h"
 #include "engine/rdma/memory_pool.h"
 #include "engine/rdma/rdma_assignment.h"
 #include "engine/rdma/rdma_config.h"
 #include "engine/rdma/rdma_env.h"
 
-#include "utils/json.hpp"
+#include "json.hpp"
 #include "logging.h"
 
 #include <condition_variable>
