+++ csrc/env.h
@@ -1,0 +1,44 @@
+#pragma once
+
+#include <cstdlib>
+#include <iostream>
+#include <sstream>
+#include <string>
+#include <vector>
+
+#include <numa.h>
+
+namespace slime {
+
+#ifndef likely
+#define likely(x) __glibc_likely(x)
+#define unlikely(x) __glibc_unlikely(x)
+#endif
+
+template<typename T>
+T get_env(const char* name, T default_value)
+{
+    const char* val = std::getenv(name);
+    if (!val)
+        return default_value;
+
+    if constexpr (std::is_same_v<T, int>) {
+        return std::stoi(val);
+    }
+    else if constexpr (std::is_same_v<T, std::vector<std::string>>) {
+        std::vector<std::string> result;
+        std::stringstream        ss(val);
+        std::string              item;
+        while (std::getline(ss, item, ',')) {
+            result.emplace_back(item);
+        }
+        return result;
+    }
+    else {
+        static_assert(sizeof(T) == 0, "Unsupported type for get_env");
+    }
+}
+
+inline const int SLIME_LOG_LEVEL = get_env<int>("SLIME_LOG_LEVEL", 0);
+
+}  // namespace slime
