+++ csrc/torch/CMakeLists.txt
@@ -3,10 +3,10 @@ slime_option(USE_MACA "USE in MACA Platform" OFF)
 
 if (USE_GLOO_BACKEND)
     set(BACKEND_SRC gloo_backend.cpp)
-    set(ENGINE_LIBRARIES _slime_gloo _slime_utils)
+    set(ENGINE_LIBRARIES _slime_gloo)
 else ()
     set(BACKEND_SRC slime_backend.cpp)
-    set(ENGINE_LIBRARIES _slime_utils _slime_engine _slime_rdma)
+    set(ENGINE_LIBRARIES _slime_engine _slime_rdma)
 endif ()
 
 pybind11_add_module(
